refactor: extract hook_audit sink from hook.py (#968)
Move the audit-log config, path resolution, and append/rotate
primitives (HookAuditConfig, load_hook_audit_config, _audit_path_for_db,
_append_audit, and their constants) into a new scipy-free
aelfrice.hook_audit module. Importing aelfrice.hook pulls in the
retrieval stack (~220ms cold, loads scipy), which is unacceptable on the
transcript logger's sub-10ms hook path; the extracted module has no such
dependency. hook.py re-exports the moved names so existing
`from aelfrice.hook import ...` callers and tests keep working; hook_tail
now imports them from hook_audit directly. The Belief-coupled record
builders stay in hook.py. No behavior change.
