+++ src/aelfrice/hook.py
@@ -40,6 +40,18 @@ from typing import IO, Any, Final, cast
 
 try:
     from aelfrice.db_paths import active_project_context, db_path
+    from aelfrice.hook_audit import (
+        AUDIT_ROTATED_SUFFIX,
+        HookAuditConfig,
+        _append_audit,
+        _audit_path_for_db,
+        _CONFIG_FILENAME,
+        load_hook_audit_config,
+    )
+    # Re-exported so existing `from aelfrice.hook import ...` callers keep
+    # working after the #968 extraction into aelfrice.hook_audit.
+    from aelfrice.hook_audit import AUDIT_DEFAULT_MAX_BYTES  # noqa: F401
+    from aelfrice.hook_audit import AUDIT_FILENAME  # noqa: F401
     from aelfrice.context_rebuilder import (
         TRIGGER_MODE_DYNAMIC,
         TRIGGER_MODE_MANUAL,
@@ -175,7 +187,6 @@ _CWD_KEY: Final[str] = "cwd"
 _UPS_SECTION: Final[str] = "user_prompt_submit_hook"
 _COLLAPSE_KEY: Final[str] = "collapse_duplicate_hashes"
 _PROMPT_SHAPE_GATE_KEY: Final[str] = "prompt_shape_gate_enabled"
-_CONFIG_FILENAME: Final[str] = ".aelfrice.toml"
 # #909: conversation-aware retrieval. The live per-prompt UPS retrieval
 # BM25s the literal prompt only; when the topic vocabulary lives in the
 # dialog history (paraphrase / pronoun / numeric reference) and not in
@@ -510,164 +521,17 @@ def _append_telemetry(
 # ---------------------------------------------------------------------------
 # Per-turn audit log (#280 mitigation 3)
 # ---------------------------------------------------------------------------
-
-AUDIT_DEFAULT_MAX_BYTES: Final[int] = 10 * 1024 * 1024
-"""Default size cap before rotation (10 MB). Overridable via .aelfrice.toml."""
+# Config, path resolution, and append/rotate primitives now live in
+# aelfrice.hook_audit (#968) so callers off the heavy retrieval import path
+# can reuse the sink; they are imported at the top of this module. The
+# Belief-coupled record builders stay below.
 
 AUDIT_PROMPT_PREFIX_CAP: Final[int] = 200
 """Maximum characters of the user prompt stored in an audit record."""
 
-AUDIT_FILENAME: Final[str] = "hook_audit.jsonl"
-"""Live audit log filename, sibling of memory.db under <git-common-dir>/aelfrice/."""
-
-AUDIT_ROTATED_SUFFIX: Final[str] = ".1"
-"""Single-slot rotation suffix. Rollover renames hook_audit.jsonl -> hook_audit.jsonl.1."""
-
-_AUDIT_SECTION: Final[str] = "hook_audit"
-_AUDIT_ENABLED_KEY: Final[str] = "enabled"
-_AUDIT_MAX_BYTES_KEY: Final[str] = "max_bytes"
-_AUDIT_ENV_DISABLE: Final[str] = "AELFRICE_HOOK_AUDIT"
-
 AUDIT_HOOK_USER_PROMPT_SUBMIT: Final[str] = "user_prompt_submit"
 AUDIT_HOOK_SESSION_START: Final[str] = "session_start"
 AUDIT_HOOK_SENTIMENT_FEEDBACK: Final[str] = "sentiment_feedback"
-
-
-@dataclass(frozen=True)
-class HookAuditConfig:
-    """Resolved configuration for the per-turn hook audit log.
-
-    `enabled` defaults True (audit-on) per #280 ratification — the surface
-    is monitored unless the operator explicitly opts out via env var or
-    TOML. `max_bytes` controls when the live file is rotated.
-    """
-
-    enabled: bool = True
-    max_bytes: int = AUDIT_DEFAULT_MAX_BYTES
-
-
-def load_hook_audit_config(
-    start: Path | None = None,
-    *,
-    env: dict[str, str] | None = None,
-    stderr: IO[str] | None = None,
-) -> HookAuditConfig:
-    """Resolve the [hook_audit] config.
-
-    Resolution order:
-    1. `AELFRICE_HOOK_AUDIT=0` env var → disabled (overrides TOML).
-    2. Walk up from `start` looking for `.aelfrice.toml`; first hit wins.
-    3. Default (enabled=True, max_bytes=AUDIT_DEFAULT_MAX_BYTES).
-
-    Missing file / missing section / malformed TOML / wrong-typed values
-    all degrade to the safe default with a stderr trace; never raises.
-    """
-    serr: IO[str] = stderr if stderr is not None else sys.stderr
-    env_map = env if env is not None else dict(os.environ)
-    env_val = env_map.get(_AUDIT_ENV_DISABLE)
-    if env_val is not None and env_val.strip() == "0":
-        return HookAuditConfig(enabled=False)
-    current = (start if start is not None else Path.cwd()).resolve()
-    seen: set[Path] = set()
-    while current not in seen:
-        seen.add(current)
-        candidate = current / _CONFIG_FILENAME
-        if candidate.is_file():
-            try:
-                raw = candidate.read_bytes()
-            except OSError as exc:
-                print(
-                    f"aelfrice hook: cannot read {candidate}: {exc}",
-                    file=serr,
-                )
-                return HookAuditConfig()
-            try:
-                parsed: dict[str, Any] = tomllib.loads(
-                    raw.decode("utf-8", errors="replace"),
-                )
-            except tomllib.TOMLDecodeError as exc:
-                print(
-                    f"aelfrice hook: malformed TOML in {candidate}: {exc}",
-                    file=serr,
-                )
-                return HookAuditConfig()
-            section_obj: Any = parsed.get(_AUDIT_SECTION, {})
-            if not isinstance(section_obj, dict):
-                return HookAuditConfig()
-            section = cast(dict[str, Any], section_obj)
-            enabled_obj: Any = section.get(_AUDIT_ENABLED_KEY, True)
-            if not isinstance(enabled_obj, bool):
-                print(
-                    f"aelfrice hook: ignoring [{_AUDIT_SECTION}] "
-                    f"{_AUDIT_ENABLED_KEY} in {candidate} (expected bool)",
-                    file=serr,
-                )
-                enabled_obj = True
-            max_bytes_obj: Any = section.get(
-                _AUDIT_MAX_BYTES_KEY, AUDIT_DEFAULT_MAX_BYTES,
-            )
-            if not isinstance(max_bytes_obj, int) or max_bytes_obj <= 0:
-                if not (
-                    isinstance(max_bytes_obj, int)
-                    and max_bytes_obj == AUDIT_DEFAULT_MAX_BYTES
-                ):
-                    print(
-                        f"aelfrice hook: ignoring [{_AUDIT_SECTION}] "
-                        f"{_AUDIT_MAX_BYTES_KEY} in {candidate} "
-                        f"(expected positive int)",
-                        file=serr,
-                    )
-                max_bytes_obj = AUDIT_DEFAULT_MAX_BYTES
-            return HookAuditConfig(
-                enabled=enabled_obj,
-                max_bytes=max_bytes_obj,
-            )
-        parent = current.parent
-        if parent == current:
-            break
-        current = parent
-    return HookAuditConfig()
-
-
-def _audit_path_for_db(db_path_val: Path) -> Path:
-    """Derive the audit log path from the DB path. Sibling of memory.db."""
-    return db_path_val.parent / AUDIT_FILENAME
-
-
-def _append_audit(
-    audit_path: Path,
-    record: dict[str, object],
-    max_bytes: int,
-    *,
-    stderr: IO[str] | None = None,
-) -> None:
-    """Append one record to the audit JSONL. Rotate if size cap exceeded.
-
-    Append-then-rotate semantics: the record always lands. If, after
-    writing, the live file exceeds `max_bytes`, it is renamed to
-    `<path>.1` (overwriting any prior `.1`) and a fresh empty file is
-    started for the next call. Single-slot rotation by spec; no archive.
-
-    Fail-soft: any I/O error is logged to stderr and swallowed.
-    """
-    try:
-        audit_path.parent.mkdir(parents=True, exist_ok=True)
-        line = json.dumps(record) + "\n"
-        with open(audit_path, "a", encoding="utf-8") as f:
-            f.write(line)
-            f.flush()
-            os.fsync(f.fileno())
-        if audit_path.stat().st_size > max_bytes:
-            rotated = audit_path.with_name(
-                audit_path.name + AUDIT_ROTATED_SUFFIX,
-            )
-            os.replace(audit_path, rotated)
-    except Exception as exc:
-        serr = stderr if stderr is not None else sys.stderr
-        print(
-            f"aelfrice: hook audit write failed (non-fatal): {exc}",
-            file=serr,
-        )
 
 
 AUDIT_BELIEF_SNIPPET_CAP: Final[int] = 120
